feat(FindObjectJobResponse): improved docs and image now optional
We've decied that imageResults will only represent AI computation. This
means HITL logic/results will not be present here. We've done this to
simplify the verdict field.

Veridct now represents whether or not (purely based on AI) we're certain
a class exists. We've fine tuned our models and chosen confidence
threshold values to ensure this is correct most of the time. Meaning
we'll try our best to give you no false positives or false negatives.

There are 3 possible values for "verdict"...

true: We're certain that this class exists
false: We're certain that this class does not exist
null: We're not certain enough to give you an answer

The other change we've made is to image in JobResults.

Image is the "chosen" image. That is to say, For this class, we've decided
that this is "the best image" to represent the class. Note that depending
if HITL is true/false, the value of image will differ slightly.

If HITL is true and an image was chosen then `image` will be the hand
chosen image.

If HITL is true and nothing was chosen, `image` will be null. This is
the case even if AI has a confidence of 100% because HITL will
supersede AI. That is to say, human intelligence is more trustworthy
than AI.

If HITL is false then 2 things can happen:

1. AI is confident enough (i.e. the confidence is above a certain
threshold). The image wth the highest confidence value above the
the threshold will be chosen.
2. AI is unsure (all confidence values are within an unsure range).
When this is the caese, no image will be chosen and image will be null.