chore(travis): updated deploy api_key
NPM decided to invalidate all auth tokens.

```
We have now invalidated all npm tokens issued before 2018-07-12 12:30 UTC, eliminating the possibility of stolen tokens being used maliciously. This is the final immediate operational action we expect to take today.

We will be conducting a forensic analysis of this incident to fully establish how many packages and users were affected, but our current belief is that it was a very small number. We will be conducting a deep audit of all the packages in the Registry to confirm this.
Posted about 6 hours ago. Jul 12, 2018 - 18:52 UTC
```

https://status.npmjs.org/incidents/dn7c1fgrr7ng